+++ packages/core/transpiler/reactivity-parser/src/parser.ts
@@ -14,6 +14,7 @@ import {
   type EnvParticle,
   type SubviewParticle,
   SwitchParticle,
+  TryParticle,
 } from "./types"
 import { type NodePath, type types as t, type traverse } from "@babel/core"
 import {
@@ -28,6 +29,7 @@ import {
   type ExpUnit,
   type SubviewUnit,
   SwitchUnit,
+  TryUnit,
 } from "@dlightjs/view-parser"
 import { DLError } from "./error"
 
@@ -98,6 +100,7 @@ export class ReactivityParser {
     if (viewUnit.type === "html") return this.parseHTML(viewUnit)
     if (viewUnit.type === "comp") return this.parseComp(viewUnit)
     if (viewUnit.type === "for") return this.parseFor(viewUnit)
+    if (viewUnit.type === "try") return this.parseTry(viewUnit)
     if (viewUnit.type === "if") return this.parseIf(viewUnit)
     if (viewUnit.type === "env") return this.parseEnv(viewUnit)
     if (viewUnit.type === "exp") return this.parseExp(viewUnit)
@@ -466,6 +469,23 @@ export class ReactivityParser {
     }
   }
 
+  // ---- @Try ----
+  /**
+   * @brief Parse a TryUnit into an TryParticle with dependencies
+   * @param tryUnit
+   * @returns TryParticle
+   */
+  private parseTry(tryUnit: TryUnit): TryParticle {
+    return {
+      type: "try",
+      children: tryUnit.children.map(this.parseViewParticle.bind(this)),
+      exception: tryUnit.exception,
+      catchChildren: tryUnit.catchChildren.map(
+        this.parseViewParticle.bind(this)
+      ),
+    }
+  }
+
   // ---- @Env ----
   /**
    * @brief Parse an EnvUnit into an EnvParticle with dependencies
@@ -570,6 +590,16 @@ export class ReactivityParser {
     dependencyIndexArr: number[]
     dependenciesNode: t.ArrayExpression
   } {
+    if (
+      this.t.isFunctionExpression(node) ||
+      this.t.isArrowFunctionExpression(node)
+    ) {
+      return {
+        dynamic: false,
+        dependencyIndexArr: [],
+        dependenciesNode: this.t.arrayExpression([]),
+      }
+    }
     // ---- Both id and prop deps need to be calculated because
     //      id is for subview update, prop is normal update
     //      in a subview, the depsNode should be both id and prop
@@ -675,12 +705,6 @@ export class ReactivityParser {
   private getPropertyDependencies(
     node: t.Expression | t.Statement
   ): [number[], t.Node[]] {
-    if (
-      this.t.isFunctionExpression(node) ||
-      this.t.isArrowFunctionExpression(node)
-    )
-      return [[], []]
-
     const deps = new Set<string>()
     const assignDeps = new Set<string>()
     const depNodes: Record<string, t.Node[]> = {}