+++ packages/core/transpiler/reactivity-parser/src/test/Dependency.test.ts
@@ -5,13 +5,13 @@ import { type CompParticle } from "../types"
 describe("Dependency", () => {
   it("should parse the correct dependency", () => {
     const viewParticles = parse("Comp(this.flag)")
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toContain(0)
   })
 
   it("should parse the correct dependency when interfacing the dependency chain", () => {
     const viewParticles = parse("Comp(this.doubleCount)")
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     const dependency = content?.dependencyIndexArr
     // ---- doubleCount depends on count, count depends on flag
     //      so doubleCount depends on flag, count and doubleCount
@@ -22,29 +22,29 @@ describe("Dependency", () => {
 
   it("should not parse the dependency if the property is not in the availableProperties", () => {
     const viewParticles = parse("Comp(this.notExist)")
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toHaveLength(0)
   })
 
   it("should not parse the dependency if the identifier is not an property of a ThisExpression", () => {
     const viewParticles = parse("Comp(count)")
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toHaveLength(0)
   })
 
   it("should not parse the dependency if the member expression is in an escaped function", () => {
     let viewParticles = parse("Comp(escape(this.flag))")
-    let content = (viewParticles[0] as CompParticle).content
+    let content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toHaveLength(0)
 
     viewParticles = parse("Comp($(this.flag))")
-    content = (viewParticles[0] as CompParticle).content
+    content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toHaveLength(0)
   })
 
   it("should not parse the dependency if the member expression is in a manual function", () => {
     const viewParticles = parse("Comp(manual(() => this.count, []))")
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toHaveLength(0)
   })
 
@@ -52,26 +52,26 @@ describe("Dependency", () => {
     const viewParticles = parse(
       "Comp(manual(() => {let a = this.count}, [this.flag]))"
     )
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toHaveLength(1)
   })
 
   it("should not parse the dependency if the member expression is the left side of an assignment expression", () => {
     const viewParticles = parse("Comp(this.flag = 1)")
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toHaveLength(0)
   })
 
   it("should not parse the dependency if the member expression is right side of an assignment expression", () => {
     const viewParticles = parse("Comp(this.flag = this.flag + 1)")
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toHaveLength(0)
   })
 
   it("should parse the dependency as identifiers", () => {
     reactivityConfig.dependencyParseType = "identifier"
     const viewParticles = parse("Comp(flag + count)")
-    const content = (viewParticles[0] as CompParticle).content
+    const content = (viewParticles[0] as CompParticle).props._$content
     expect(content?.dependencyIndexArr).toContain(
       availableProperties.indexOf("flag")
     )