+++ packages/core/transpiler/view-generator/src/HelperGenerators/BaseGenerator.ts
@@ -190,7 +190,7 @@ export default class BaseGenerator {
   }
 
   get updateParams() {
-    return [this.t.identifier("$changed"), this.t.identifier("$key")]
+    return [this.t.identifier("$changed")]
   }
 
   /**